+++ src/components/researchPortal/ResearchExpressionPlot.vue
@@ -34,9 +34,13 @@
 				</div>
 			</div>
 		</div>
-		<div id="multi-chart">
-			<p>Loading...</p>
-		</div>
+		<research-gene-expression-plot
+			:flatBoth="flatBoth"
+			:colorMap="colorMap"
+			:keyField="keyField"
+			:logScale="logScale"
+		>
+		</research-gene-expression-plot>
 		<research-expression-table
 			:tableData="tableData">
 		</research-expression-table>
@@ -48,13 +52,12 @@ import * as d3 from "d3";
 import uiUtils from "@/utils/uiUtils";
 import colors from "@/utils/colors";
 import Formatters from "@/utils/formatters";
-import ResearchExpressionTable from "@/components/researchPortal/ResearchExpressionTable.vue"
+import ResearchExpressionTable from "@/components/researchPortal/ResearchExpressionTable.vue";
+import ResearchGeneExpressionPlot from "./ResearchGeneExpressionPlot.vue";
 export default Vue.component("ResearchExpressionPlot", {
 	props: ["rawData", "filter", "plotByField", "skipSort"],
 	data() {
 		return {
-			chart: null,
-			chartWidth: null,
 			logScale: true,
 			keyField: this.$props.plotByField,
 			keyFieldList: [],
@@ -109,21 +112,13 @@ export default Vue.component("ResearchExpressionPlot", {
 		},
 	},
 	mounted() {
-		this.chart = document.getElementById("multi-chart");
-		this.chartWidth = this.chart.clientWidth;
-		addEventListener("resize", (event) => {
-			this.chartWidth = this.chart.clientWidth;
-			this.displayResults();
-		});
 		this.processData();
-		this.displayResults();
 	},
 	methods: {
 		...uiUtils,
 
 		applyFilter() {
 			this.processData();
-			this.displayResults();
 		},
 
 		tpmFormat(value) {
@@ -196,321 +191,8 @@ export default Vue.component("ResearchExpressionPlot", {
 				}
 			}
 			this.processedData = processedData;
-
 			this.flatBoth = flatBoth;
 			this.mapColors();
-		},
-		displayResults() {
-			let colorMap = this.colorMap;
-			let flatData = this.flatBoth;
-			let dotBoxHalfWidth = 6;
-
-			let tpmField = this.logScale ? "log" : "linear";
-			let margin = {
-					top: 10,
-					right: 30,
-					bottom: this.getBottomMargin(flatData, this.keyField),
-					left: 40,
-				},
-				width = this.chartWidth - margin.left - margin.right,
-				height = 400 - margin.top - margin.bottom;
-			this.chart.innerHTML = "";
-			let svg = d3
-				.select("#multi-chart")
-				.append("svg")
-				.attr("width", width + margin.left + margin.right)
-				.attr("height", height + margin.top + margin.bottom)
-				.append("g")
-				.attr(
-					"transform",
-					"translate(" + margin.left + "," + margin.top + ")"
-				)
-				.on("mouseleave", (d) => svg.selectAll("circle").remove());
-			let tooltip = d3
-				.select("#multi-chart")
-				.append("div")
-				.style("opacity", 0)
-				.attr("class", "tooltip")
-				.style("background-color", "white")
-				.style("border", "2px solid gray")
-				.style("padding", "5px")
-				.style("border-radius", "5px")
-				.style("font-size", "smaller");
-
-			let x = d3
-				.scaleBand()
-				.range([0, width])
-				.domain(flatData.map((entry) => entry[this.keyField]))
-				.padding(0.05);
-
-			svg.append("g")
-				.attr("transform", `translate(0,${height})`)
-				.call(d3.axisBottom(x))
-				.selectAll("text")
-				.style("text-anchor", "start")
-				.style("font-size", "13px")
-				.attr("transform", "rotate(45)");
-
-			let maxVal = flatData
-				.map((g) => g[tpmField])
-				.reduce((prev, next) => (prev > next ? prev : next), 0);
-			let y = d3.scaleLinear().domain([0, maxVal]).range([height, 0]);
-
-			svg.append("g").call(d3.axisLeft(y));
-
-			let histogram = d3
-				.histogram()
-				.domain(y.domain())
-				.thresholds(y.ticks(100))
-				.value((d) => d);
-			let sumstat = d3
-				.nest()
-				.key((d) => d[this.keyField])
-				.rollup((d) => {
-					let input = d.map((g) => g[tpmField]);
-					let bins = histogram(input);
-					return bins;
-				})
-				.entries(flatData);
-
-			//Maximum number of entries in a bin.
-			let maxNum = 0;
-			for (let i in sumstat) {
-				let allBins = sumstat[i].value;
-				let lengths = allBins.map((a) => a.length);
-				let longest = d3.max(lengths);
-				if (longest > maxNum) {
-					maxNum = longest;
-				}
-			}
-			let xNum = d3
-				.scaleLinear()
-				.range([0, x.bandwidth()])
-				.domain([-maxNum, maxNum]);
-
-			let colorIndex = 0;
-			let violinIndex = 0;
-			let mouseover = (d) => {
-				svg.selectAll(".violin").style("opacity", 1);
-				let violinNumber = this.keyFieldList.indexOf(d.key);
-				svg.selectAll(`.violin_${violinNumber}`).style("opacity", 0.25);
-				svg.selectAll("circle").remove();
-				svg.selectAll("indPoints")
-					.data(flatData.filter((entry) => entry[this.keyField] === d.key))
-					.enter()
-					.append("circle")
-					.attr("class", (g) => g.dataset)
-					.attr("cx", (g) => {
-						let dx =
-							offset -
-							2 * dotBoxHalfWidth +
-							g.noise * dotBoxHalfWidth * 4;
-						return x(d.key) + dx;
-					})
-					.attr("cy", (g) => y(g[tpmField]))
-					.attr("r", 2)
-					.attr("fill", "none")
-					.attr("stroke", "lightgray")
-					.on("mouseover", hoverDot)
-					.on("mouseleave", hideTooltip);
-			};
-			let redrawHoverDots = (g) => {
-				let hoverItem = g[this.keyField];
-				let hoverDataset = g.dataset;
-				let hoverColor = `${colorMap[g[this.keyField]]}`;
-				svg.selectAll("indPoints")
-					.data(flatData.filter((entry) =>
-						entry[this.keyField] == hoverItem && entry.dataset == hoverDataset))
-					.enter()
-					.append("circle")
-					.attr("class", (j) => j.dataset)
-					.attr("cx", (j) => {
-						let dx =
-							offset -
-							2 * dotBoxHalfWidth +
-							j.noise * dotBoxHalfWidth * 4;
-						return x(g[this.keyField]) + dx;
-					})
-					.attr("cy", (j) => y(j[tpmField]))
-					.attr("r", 2)
-					.attr("fill", "none")
-					.attr("stroke", hoverColor)
-					.on("mouseover", hoverDot)
-					.on("mouseleave", hideTooltip);
-			};
-			let redrawNonHoverDots = (g) => {
-				let hoverItem = g[this.keyField];
-				let hoverDataset = g.dataset;
-				svg.selectAll("indPoints")
-					.data(flatData.filter((entry) =>
-						entry[this.keyField] === hoverItem && entry.dataset != hoverDataset))
-					.enter()
-					.append("circle")
-					.attr("class", (j) => j.dataset)
-					.attr("cx", (j) => {
-						let dx =
-							offset -
-							2 * dotBoxHalfWidth +
-							j.noise * dotBoxHalfWidth * 4;
-						return x(g[this.keyField]) + dx;
-					})
-					.attr("cy", (j) => y(j[tpmField]))
-					.attr("r", 2)
-					.attr("fill", "none")
-					.attr("stroke", "lightgray")
-					.on("mouseover", hoverDot)
-					.on("mouseleave", hideTooltip);
-			};
-			let hoverDot = (g) => {
-				let xcoord = `${d3.event.layerX + 35}px`;
-				let ycoord = `${d3.event.layerY}px`;
-				svg.selectAll("circle").remove();
-				redrawNonHoverDots(g);
-				redrawHoverDots(g);
-				// Tooltip content
-				let tooltipContent = `Biosample: ${g.biosample}`;
-				tooltipContent = tooltipContent.concat(
-					`<span>Dataset: ${g.dataset}</span>`
-				);
-				tooltip
-					.style("opacity", 1)
-					.html(tooltipContent)
-					.style("left", xcoord)
-					.style("top", ycoord);
-			};
-			let hideTooltip = (g) => {
-				tooltip.style("opacity", 0);
-			};
-			svg.selectAll("myViolin")
-				.data(sumstat)
-				.enter()
-				.append("g")
-				.attr("transform", (d) => `translate(${x(d.key)},0)`)
-				.append("path")
-				.datum((d) => d.value)
-				.attr("class", (d) => {
-					let classString = `violin violin_${violinIndex}`;
-					violinIndex++;
-					return classString;
-				})
-				.style("stroke", "none")
-				.style("fill", (d) => {
-					// I don't like reinventing the wheel, but I cannot
-					// figure out how to access the key attribute as text
-					let color = colors[colorIndex];
-					colorIndex++;
-					if (colorIndex >= colors.length) {
-						colorIndex = 0;
-					}
-					return color;
-				})
-				.attr(
-					"d",
-					d3
-						.area()
-						.x0((d) => xNum(-d.length))
-						.x1((d) => xNum(d.length))
-						.y((d) => y(d.x0))
-						.curve(d3.curveCatmullRom)
-				)
-				.on("mouseover", mouseover);
-			let numberViolins = 0;
-			let sumstatBox = d3
-				.nest()
-				.key((d) => d[this.keyField])
-				.rollup((d) => {
-					numberViolins++;
-					let sortedData = d
-						.map((g) => g[tpmField])
-						.sort(d3.ascending);
-					let q1 = d3.quantile(sortedData, 0.25);
-					let median = d3.quantile(sortedData, 0.5);
-					let q3 = d3.quantile(sortedData, 0.75);
-					let interQuantileRange = q3 - q1;
-					let min = sortedData[0];
-					let max = sortedData[sortedData.length - 1];
-					let boxplotEntry = {
-						"Q1 TPM": q1,
-						"Median TPM": median,
-						"Q3 TPM": q3,
-						interQuantileRange: interQuantileRange,
-						"Min TPM": min,
-						"Max TPM": max,
-						"Total samples": sortedData.length,
-					};
-					return boxplotEntry;
-				})
-				.entries(flatData);
-			let offset = width / (2 * numberViolins);
-			// Boxplots top quartile
-			svg.selectAll("vertLines")
-				.data(sumstatBox)
-				.enter()
-				.append("line")
-				.attr("x1", x(0))
-				.attr("x2", x(0))
-				.attr("y1", (d) => y(d.value["Q3 TPM"]))
-				.attr("y2", (d) => y(d.value["Max TPM"]))
-				.attr("stroke", "black")
-				.style("opacity", 0.5)
-				.style("width", 30)
-				.attr("transform", (d) => `translate(${x(d.key) + offset},0)`)
-				.on("mouseover", mouseover);
-			// Boxplots bottom quartile
-			svg.selectAll("vertLines")
-				.data(sumstatBox)
-				.enter()
-				.append("line")
-				.attr("x1", x(0))
-				.attr("x2", x(0))
-				.attr("y1", (d) => y(d.value["Min TPM"]))
-				.attr("y2", (d) => y(d.value["Q1 TPM"]))
-				.attr("stroke", "black")
-				.style("opacity", 0.5)
-				.style("width", 30)
-				.attr("transform", (d) => `translate(${x(d.key) + offset},0)`)
-				.on("mouseover", mouseover);
-			let boxHalfWidth = 3;
-			svg.selectAll("boxes")
-				.data(sumstatBox)
-				.enter()
-				.append("rect")
-				.attr("x", (d) => x(d.key) + offset - boxHalfWidth)
-				.attr("y", (d) => y(d.value["Q3 TPM"]))
-				.attr(
-					"height",
-					(d) => y(d.value["Q1 TPM"]) - y(d.value["Q3 TPM"])
-				)
-				.attr("width", boxHalfWidth * 2)
-				.attr("stroke", "black")
-				.style("fill", "white")
-				.style("opacity", 0.5)
-				.on("mouseover", mouseover);
-
-			// Packaging data for export at the same time.
-			svg.selectAll("zoneBoxes")
-				.data(sumstatBox)
-				.enter()
-				.append("rect")
-				.attr("x", (d) => x(d.key))
-				.attr("y", (d) => y(maxVal))
-				.attr("height", (d) => y(-maxVal) - y(0))
-				.attr("width", offset * 2)
-				.attr("stroke", "none")
-				.style("fill", "white")
-				.style("opacity", 0)
-				.on("mouseover", mouseover);
-			svg.selectAll("medianLines")
-				.data(sumstatBox)
-				.enter()
-				.append("line")
-				.attr("x1", (d) => x(d.key) + offset - boxHalfWidth)
-				.attr("x2", (d) => x(d.key) + offset + boxHalfWidth)
-				.attr("y1", (d) => y(d.value["Median TPM"]))
-				.attr("y2", (d) => y(d.value["Median TPM"]))
-				.attr("stroke", "#99999999")
-				.style("width", 50)
-				.on("mouseover", mouseover);
 		},
 		mapColors() {
 			let colorMap = {};
@@ -526,44 +208,6 @@ export default Vue.component("ResearchExpressionPlot", {
 			});
 			this.colorMap = colorMap;
 		},
-		getBottomMargin(data, labelField) {
-			let longestLabel = data
-				.map((item) => item[labelField].length)
-				.reduce((prev, next) => (prev > next ? prev : next), 0);
-			let margin = longestLabel < 10 ? 65 : (65 * longestLabel) / 10;
-			return margin;
-		},
 	},
 });
 </script>
-<style>
-div {
-	display: block;
-}
-.chart-wrapper {
-	display: block;
-}
-.chart-wrapper > label {
-	/*font-size: smaller;*/
-	padding: 10px;
-}
-.chart {
-	flex: 1;
-}
-.all-charts {
-	display: flex;
-	margin: 20px;
-}
-#select-gene {
-	width: 200px;
-}
-.dataset-subtable {
-	font-size: smaller;
-}
-#big-table > thead > tr > th {
-	color: #007bff;
-}
-.tooltip span {
-	display: block;
-}
-</style>
